Refactor out to datasette.utils.make_slot_function()
